+++ go-calendar/server/launcher_test.go
@@ -1,4 +1,4 @@
-package services
+package server
 
 import (
 	"net/http"
